+++ CHANGELOG.md
@@ -21,3 +21,7 @@
 
 - Add catalog endpoint
 - Add collection endpoint
+
+## [1.0.4] - 2026-03-26
+
+- Add product variant endpoint